Move code to more proper places
